requirements.txt
geopandas
matplotlib
Pillow
faiss-cpu
transformers
torch